correct merge conflicts

+++ core/engine/src/builtins/json/mod.rs
@@ -1119,7 +1119,7 @@ impl Json {
         let len = value.length_of_array_like(context)?;
 
         // 5. Let partial be a new empty List.
-        let mut partial = Vec::with_capacity(len as usize);
+        let mut partial = Vec::with_capacity(len);
 
         // 7. Let index be 0.
         let mut index = 0;
